requirements.txt
Django
djangorestframework
textblob
mysqlclient
drf_yasg